Set the static lib as the default lib for the unittests.
The unittests will fail if they are linked with the shared lib, and the
shared lib is not in a path where ld will look for it when running the
executable.

Alternatively we could use the -rpath flag to add the directory
containing the newly build shared lib to the ld search path for the
unittests.